Moved data into package
