+++ 3_Python/package/dnn/snn/net_utility.py
@@ -5,6 +5,7 @@ from snntorch import utils
 
 
 def save_net(net, path: str) -> None:
+    """Loading """
     path2model = abspath("../scr_ai")
     rel_path = join(path2model, path)
     utils.reset(net)
@@ -17,11 +18,13 @@ def save_net(net, path: str) -> None:
 
 
 def load_net(path, file_name: str):
+    """Loading a """
     path2model = os.path.abspath("../scr_ai")
     rel_path = os.path.join(path2model, path)
     com_path = os.path.join(rel_path, file_name+".pt")
     if not os.path.exists(com_path):
         raise Exception(f'model does not exist: {com_path}')
+        return 0
     else:
         path = os.path.join('../scr_ai', path)
         return load(os.path.join(path, file_name+".pt"))
